Make menu highlighting work
The navbar now keeps up with the section

I'm using JS to do this, and I'm not entirely happy with the
way it works because it flickers when switching pages. I need to figure
a way to make the go template set the correct section even for an index.

- also got indexes working
- rename projects -> project

TODO extract the li template into a fragment which would make indexes
easier to re-style. Starting to see some of the downsides to static
blogging. Very much you have to store the state in the page as hardcoded
values or markup. Trading one set of problems for a different set is
always interesting.